Struct rabbit::Rabbit

pub struct Rabbit {
    // some fields omitted
}

Methods

impl Rabbit

fn new(key: &Key) -> Rabbit

Setupы given key on an empty rabbit state.

fn new_iv(key: &Key, iv: &InitVec) -> Rabbit

Setupы given key on an empty rabbit state, then setupы initialization vector iv on it.

fn reset(&mut self)

Restores master state.

fn reinit(&mut self, iv: &InitVec)

Restores master state, than setups initialization vector iv on it.

fn encrypt(&mut self, data: &[u8], buf: &mut [u8])

Encrypts and writes bytes of data to buf. Asserts that buf.len() >= data.len().

fn decrypt(&mut self, data: &[u8], buf: &mut [u8])

Decrypts and writes bytes of data to buf. Asserts that buf.len() >= data.len().

fn encrypt_inplace(&mut self, data: &mut [u8])

Encrypts bytes of data inplace.

fn decrypt_inplace(&mut self, data: &mut [u8])

Decrypts bytes of data inplace.